Fix MP3 playback: register Libmpg123 factory in MediaFactory
The mpeg_audio format was defined but the registerFormatInternal call
was commented out with a TODO note. This caused MP3 files to be detected
correctly but fail with 'unsupported format' when trying to create a
stream.

Uncommented and updated the registration to use the correct Libmpg123
class from PsyMP3::Codec::MP3 namespace.